Feature/author kh

+++ src/main/java/library/repository/BookRepository.java
@@ -1,10 +1,10 @@
 package library.repository;
 
-import library.entity.Book;
+import library.model.book.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface BookRepository extends JpaRepository<Book, Long> {
-    List<Book> findAllByAuthor_Id(Integer id);
+    List<Book> findAllByAuthor_Id(Long id);
 }
